platform/services/auth-service/pyproject.toml
[tool.poetry]
name = "auth-service"
version = "1.0.0"
description = "Authentication service for ALGORITHMIC ARTS platform"
authors = ["Alex <alex@algorithmic-arts.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
pydantic = {extras = ["email"], version = "^2.10.0"}
pydantic-settings = "^2.6.0"
redis = {extras = ["hiredis"], version = "^5.2.0"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyotp = "^2.9.0"
qrcode = "^8.0"
python-multipart = "^0.0.18"
httpx = "^0.28.0"
aiokafka = "^0.11.0"
structlog = "^24.4.0"
prometheus-fastapi-instrumentator = "^7.0.0"
slowapi = "^0.1.9"

[tool.poetry.dev-dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
httpx = "^0.28.0"
factory-boy = "^3.3.0"
faker = "^33.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"